Added ProgressBar

+++ mainwindow.h
@@ -39,7 +39,7 @@ private slots:
 	void on_toEMU_clicked();
 
 private:
-	void copy_file_skip(const QString &in, const QString &out, unsigned int skip);
+	void copy_file_skip(const QString &in, const QString &out, qint64 skip);
 
 	Ui::MainWindow *ui;
 };
